refactor(llama): split inference session out
