+++ llama-rs/src/lib.rs
@@ -1,37 +1,30 @@
 #![deny(missing_docs)]
 //! LLaMA-rs is a Rust port of the llama.cpp project. This allows running inference for Facebook's LLaMA model on a CPU with good performance using full precision, f16 or 4-bit quantized versions of the model.
 
-use core::slice;
 use std::{
     collections::HashMap,
     fmt::Display,
     io::{BufRead, Read, Seek, SeekFrom},
     path::{Path, PathBuf},
     str::FromStr,
-    time,
 };
 
 use serde::Deserialize;
 use thiserror::Error;
 
-use partial_sort::PartialSort;
-use rand::{distributions::WeightedIndex, prelude::Distribution};
-
-pub use ggml::Type as ElementType;
-
 #[cfg(feature = "convert")]
 pub mod convert;
 
+mod inference_session;
 mod util;
+
+pub use ggml::Type as ElementType;
+pub use inference_session::{
+    InferenceSession, InferenceSessionParameters, InferenceSnapshot, ModelKVMemoryType,
+};
 
 /// The end of text token.
 pub const EOT_TOKEN_ID: TokenId = 2; // Hardcoded (for now?)
-
-// The size of a scratch buffer used for inference. This is used for temporary
-// storage of intermediate results during inference.
-//
-// The specific value was copied from `llama.cpp`.
-const SCRATCH_SIZE: usize = 512 * 1024 * 1024;
 
 /// The hyperparameters of the model.
 #[derive(Debug, Default, PartialEq, Eq, PartialOrd, Ord, Deserialize)]
@@ -79,176 +72,6 @@ pub struct Model {
 
     // Must be kept alive for the model
     _context: ggml::Context,
-}
-
-/// An inference session represents the state of the text generation. This holds
-/// the full context window, as long as several additional parameters used
-/// during sampling.
-pub struct InferenceSession {
-    // Must be kept alive for the model
-    _session_ctx: ggml::Context,
-
-    // Original size of the memory used to create this context.
-    memory_size: usize,
-
-    // Parameters for the session.
-    params: InferenceSessionParameters,
-
-    memory_k: ggml::Tensor,
-    memory_v: ggml::Tensor,
-
-    /// How many tokens have been fed into the model's working memory so far.
-    n_past: usize,
-
-    /// How much memory is required per token for the temporary context used
-    /// during inference.
-    mem_per_token: usize,
-
-    /// All tokens generated by this inference session
-    tokens: Vec<TokenId>,
-
-    /// The logits that were last predicted by the network. Zeroed out otherwise.
-    last_logits: Vec<f32>,
-
-    /// Scratch buffers used during inference.
-    ///
-    /// The number of scratch buffers was copied from `llama.cpp`.
-    /// There is no specific reason for this number, but one is insufficient.
-    scratch: [ggml::Buffer; 2],
-}
-impl InferenceSession {
-    fn repetition_penalty_tokens(&self) -> &[TokenId] {
-        &self.tokens[self
-            .tokens
-            .len()
-            .saturating_sub(self.params.repetition_penalty_last_n)..]
-    }
-}
-impl Clone for InferenceSession {
-    fn clone(&self) -> Self {
-        let context = ggml::Context::init(self.memory_size);
-        let memory_k = context.new_tensor_1d(self.memory_k.get_type(), self.memory_k.nelements());
-        let memory_v = context.new_tensor_1d(self.memory_v.get_type(), self.memory_v.nelements());
-
-        Self {
-            _session_ctx: context,
-            memory_size: self.memory_size,
-            params: self.params,
-            memory_k,
-            memory_v,
-            n_past: self.n_past,
-            mem_per_token: self.mem_per_token,
-            tokens: self.tokens.clone(),
-            last_logits: self.last_logits.clone(),
-            scratch: inference_session_scratch_buffers(),
-        }
-    }
-}
-
-fn inference_session_scratch_buffers() -> [ggml::Buffer; 2] {
-    [
-        ggml::Buffer::new(SCRATCH_SIZE),
-        ggml::Buffer::new(SCRATCH_SIZE),
-    ]
-}
-
-#[derive(serde::Serialize, Clone, PartialEq)]
-/// A serializable snapshot of the inference process.
-/// Can be created by calling [InferenceSession::get_snapshot].
-///
-/// If serializing, ensure that your serializer is binary-efficient.
-/// This type contains a large array of bytes; traditional textual serializers
-/// are likely to serialize this as an array of numbers at extreme cost.
-// Keep in sync with [InferenceSession] and [InferenceSnapshot].
-pub struct InferenceSnapshotRef<'a> {
-    /// How many tokens have been stored in the memory so far.
-    pub npast: usize,
-    /// Parameters associated with the saved inference session.
-    pub session_params: InferenceSessionParameters,
-    /// All tokens generated by this inference session
-    pub tokens: Vec<TokenId>,
-    /// The vector of logits that was produced after the last inference
-    pub logits: Vec<f32>,
-    /// The contents of the 'key' memory tensor
-    #[serde(with = "serde_bytes")]
-    pub memory_k: &'a [u8],
-    /// The contents of the 'value' memory tensor
-    #[serde(with = "serde_bytes")]
-    pub memory_v: &'a [u8],
-}
-impl InferenceSnapshotRef<'_> {
-    /// Creates an owned [InferenceSnapshot] from this [InferenceSnapshotRef].
-    ///
-    /// The [ToOwned] trait is not used due to its blanket implementation for all [Clone] types.
-    pub fn to_owned(&self) -> InferenceSnapshot {
-        InferenceSnapshot {
-            npast: self.npast,
-            session_params: self.session_params,
-            tokens: self.tokens.clone(),
-            last_logits: self.logits.clone(),
-            memory_k: self.memory_k.to_vec(),
-            memory_v: self.memory_v.to_vec(),
-        }
-    }
-}
-
-/// A serializable snapshot of the inference process. Can be restored by calling
-/// [Model::session_from_snapshot].
-#[derive(serde::Deserialize, Clone, PartialEq)]
-// Keep in sync with [InferenceSession] and [InferenceSnapshotRef].
-pub struct InferenceSnapshot {
-    /// How many tokens have been stored in the memory so far.
-    pub npast: usize,
-    /// Parameters associated with the saved inference session.
-    pub session_params: InferenceSessionParameters,
-    /// All tokens generated by this inference session
-    pub tokens: Vec<TokenId>,
-    /// The vector of logits that was produced after the last inference
-    pub last_logits: Vec<f32>,
-    /// The contents of the 'key' memory tensor
-    #[serde(with = "serde_bytes")]
-    pub memory_k: Vec<u8>,
-    /// The contents of the 'value' memory tensor
-    #[serde(with = "serde_bytes")]
-    pub memory_v: Vec<u8>,
-}
-
-/// Allowed types for the model memory K/V tensors.
-#[derive(Clone, Copy, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum ModelKVMemoryType {
-    /// 16-bit float.
-    Float16,
-    /// 32-bit float.
-    Float32,
-}
-impl From<ModelKVMemoryType> for ggml::Type {
-    fn from(value: ModelKVMemoryType) -> Self {
-        match value {
-            ModelKVMemoryType::Float16 => ggml::Type::F16,
-            ModelKVMemoryType::Float32 => ggml::Type::F32,
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
-/// Parameters for an inference session.
-pub struct InferenceSessionParameters {
-    /// The number of tokens to consider for the repetition penalty.
-    pub repetition_penalty_last_n: usize,
-    /// The type of the memory K tensor.
-    pub memory_k_type: ModelKVMemoryType,
-    /// The type of the memory V tensor.
-    pub memory_v_type: ModelKVMemoryType,
-}
-
-impl Default for InferenceSessionParameters {
-    fn default() -> Self {
-        Self {
-            repetition_penalty_last_n: 512,
-            memory_k_type: ModelKVMemoryType::Float32,
-            memory_v_type: ModelKVMemoryType::Float32,
-        }
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -1141,7 +964,7 @@ impl Model {
             mem_per_token: 0,
             tokens: vec![],
             last_logits: vec![0.0; n_vocab],
-            scratch: inference_session_scratch_buffers(),
+            scratch: inference_session::scratch_buffers(),
         }
     }
 
@@ -1467,266 +1290,6 @@ impl Model {
         session.last_logits = snapshot.last_logits;
 
         Ok(session)
-    }
-}
-
-impl InferenceSession {
-    /// Feed a prompt to the model for this session.
-    pub fn feed_prompt<E: std::error::Error + 'static>(
-        &mut self,
-        model: &Model,
-        vocab: &Vocabulary,
-        params: &InferenceParameters,
-        prompt: &str,
-        mut callback: impl FnMut(&[u8]) -> Result<(), E>,
-    ) -> Result<(), InferenceError> {
-        let beginning_of_sentence = self.n_past == 0;
-        let prompt_tokens: Vec<TokenId> = vocab
-            .tokenize(prompt, beginning_of_sentence)?
-            .iter()
-            .map(|(_, tok)| *tok)
-            .collect();
-
-        if self.n_past + prompt_tokens.len() >= model.hparams.n_ctx {
-            return Err(InferenceError::ContextFull);
-        }
-
-        for batch in prompt_tokens.chunks(params.n_batch) {
-            model.evaluate(self, params, batch, &mut EvaluateOutputRequest::default());
-            for &tk in batch {
-                // NOTE: No string ever tokenizes to the end of sentence. So we
-                // can just return the id here.
-                if let Err(e) = callback(vocab.token(tk as usize)) {
-                    return Err(InferenceError::UserCallback(Box::new(e)));
-                }
-
-                // Update the tokens for this session
-                self.tokens.push(tk);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Infer the next token for this session.
-    pub fn infer_next_token<'v>(
-        &mut self,
-        model: &Model,
-        vocab: &'v Vocabulary,
-        params: &InferenceParameters,
-        rng: &mut impl rand::Rng,
-    ) -> Result<&'v [u8], InferenceError> {
-        if self.n_past + 1 >= model.hparams.n_ctx {
-            return Err(InferenceError::ContextFull);
-        }
-
-        // First, sample the next token, using the stored last_logits;
-        let next_token = self.sample_top_p_top_k(params, rng);
-
-        // Update the tokens for this session
-        self.tokens.push(next_token);
-
-        // Then, evaluate the network again to compute the new last_logits
-        model.evaluate(
-            self,
-            params,
-            &[next_token],
-            &mut EvaluateOutputRequest::default(),
-        );
-
-        // Return the next token
-        if next_token as TokenId == EOT_TOKEN_ID {
-            Err(InferenceError::EndOfText)
-        } else {
-            Ok(vocab.token(next_token as usize))
-        }
-    }
-
-    // todo: see if we can reduce the arguments here somehow - consolidate model and vocab maybe?
-    /// Helper function to run inference with this session and the given model and vocabulary.
-    /// The `callback` is called with each new token until inference is complete.
-    ///
-    /// If `params.play_back_previous_tokens` is specified, this will "play back" all existing tokens in the session.
-    #[allow(clippy::too_many_arguments)]
-    pub fn inference_with_prompt<E: std::error::Error + 'static>(
-        &mut self,
-        model: &Model,
-        vocab: &Vocabulary,
-        params: &InferenceParameters,
-        prompt: &str,
-        maximum_token_count: Option<usize>,
-        rng: &mut impl rand::Rng,
-        mut callback: impl FnMut(&str) -> Result<(), E>,
-    ) -> Result<InferenceStats, InferenceError> {
-        let maximum_token_count = maximum_token_count.unwrap_or(usize::MAX);
-        if params.play_back_previous_tokens {
-            // "Play back" the existing tokens, so that loading from an inference snapshot works
-            // as expected.
-            let mut token_utf8_buf = TokenUtf8Buffer::new();
-            for token_id in &self.tokens {
-                // Buffer the token until it's valid UTF-8, then call the callback.
-                if let Some(tokens) = token_utf8_buf.push(vocab.token(*token_id as usize)) {
-                    if let Err(e) = callback(&tokens) {
-                        return Err(InferenceError::UserCallback(Box::new(e)));
-                    }
-                }
-            }
-        }
-
-        let mut stats = InferenceStats::default();
-
-        let start_at = time::SystemTime::now();
-
-        // Feed the initial prompt through the transformer, to update its
-        // context window with new data.
-        self.feed_prompt(
-            model,
-            vocab,
-            params,
-            prompt,
-            TokenUtf8Buffer::adapt_callback(&mut callback),
-        )?;
-        stats.feed_prompt_duration = start_at.elapsed().unwrap();
-        stats.prompt_tokens = self.n_past;
-
-        // After the prompt is consumed, sample tokens by repeatedly calling
-        // `infer_next_token`. We generate tokens until the model returns an
-        // EndOfText token, or we run out of space in the context window,
-        // or we reach the specified limit.
-        let mut tokens_processed = 0;
-        let mut token_utf8_buf = TokenUtf8Buffer::new();
-        while tokens_processed < maximum_token_count {
-            let token = match self.infer_next_token(model, vocab, params, rng) {
-                Ok(token) => token,
-                Err(InferenceError::EndOfText) => break,
-                Err(e) => return Err(e),
-            };
-
-            // Buffer the token until it's valid UTF-8, then call the callback.
-            if let Some(tokens) = token_utf8_buf.push(token) {
-                if let Err(e) = callback(&tokens) {
-                    return Err(InferenceError::UserCallback(Box::new(e)));
-                }
-            }
-
-            tokens_processed += 1;
-        }
-        stats.predict_duration = start_at.elapsed().unwrap();
-        stats.predict_tokens = self.n_past;
-
-        Ok(stats)
-    }
-
-    /// Sample a token using Top-P/Top-K sampling and the last logits from this session.
-    pub fn sample_top_p_top_k(
-        &self,
-        params: &InferenceParameters,
-        rng: &mut impl rand::Rng,
-    ) -> TokenId {
-        let logits = &self.last_logits;
-        let n_logits = logits.len();
-        let mut logits_id = Vec::<(f32, TokenId)>::with_capacity(n_logits);
-
-        {
-            let scale = 1.0 / params.temperature;
-            for (i, &logit) in logits.iter().enumerate() {
-                let tid = i as TokenId;
-
-                let val = if let Some(logit_override) = params.bias_tokens.get(tid) {
-                    logit_override
-                } else if self.repetition_penalty_tokens().contains(&(i as TokenId)) {
-                    // repetition penalty from CTRL paper (https://arxiv.org/abs/1909.05858)
-                    // credit https://github.com/facebookresearch/llama/compare/main...shawwn:llama:main
-
-                    // if score < 0 then repetition penalty has to multiplied to reduce the previous token probability
-                    if logits[i] < 0.0 {
-                        logit * scale * params.repeat_penalty
-                    } else {
-                        logit * scale / params.repeat_penalty
-                    }
-                } else {
-                    logit * scale
-                };
-                logits_id.push((val, tid));
-            }
-        }
-
-        // find the top K tokens
-        {
-            logits_id.partial_sort(params.top_k, |a, b| {
-                // Sort descending
-                b.0.total_cmp(&a.0)
-            });
-            logits_id.truncate(params.top_k);
-        }
-
-        let maxl = logits_id
-            .iter()
-            .map(|x| x.0)
-            .max_by(f32::total_cmp)
-            .unwrap();
-
-        // compute probs for the top K tokens
-        let mut probs: Vec<f32> = logits_id
-            .iter()
-            .copied()
-            .map(|(k, _)| (k - maxl).exp())
-            .collect();
-        let sum: f32 = probs.iter().copied().sum();
-
-        // Normalize the probs
-        for p in probs.iter_mut() {
-            *p /= sum;
-        }
-
-        // Top p sampling
-        if params.top_p < 1.0 {
-            let mut cumsum = 0.0;
-            for i in 0..probs.len() {
-                cumsum += probs[i];
-                if cumsum >= params.top_p {
-                    probs.truncate(i + 1);
-                    logits_id.truncate(i + 1);
-                    break;
-                }
-            }
-
-            cumsum = 1.0 / cumsum;
-            for p in probs.iter_mut() {
-                *p *= cumsum;
-            }
-        }
-
-        let dist = WeightedIndex::new(&probs).expect("WeightedIndex error");
-        let idx = dist.sample(rng);
-
-        logits_id[idx].1
-    }
-
-    /// Obtains a serializable snapshot of the current inference status. This
-    /// can be used to cache the state of the model and store them into a file.
-    ///
-    /// # Safety
-    ///
-    /// This function provides raw access to the underlying memory owned by the
-    /// ggml context. While the provided `InferenceSnapshotRef` object is alive,
-    /// no other methods for this model object should be called.
-    pub unsafe fn get_snapshot(&mut self) -> InferenceSnapshotRef<'_> {
-        let memory_k = unsafe {
-            slice::from_raw_parts(self.memory_k.data() as *mut u8, self.memory_k.nbytes())
-        };
-        let memory_v = unsafe {
-            slice::from_raw_parts(self.memory_v.data() as *mut u8, self.memory_v.nbytes())
-        };
-
-        InferenceSnapshotRef {
-            npast: self.n_past,
-            session_params: self.params,
-            tokens: self.tokens.clone(),
-            logits: self.last_logits.clone(),
-            memory_k,
-            memory_v,
-        }
     }
 }
